requirements.txt
streamlit==1.35.0
langgraph
langchain-google-genai
python-dotenv==1.0.0
pypdf==4.2.0
faiss-cpu==1.8.0
